Implement new Element interface rendering system
This should be much better than the previous system and works as follows:

  * If any change occurs on an Element, whether the root window or a deeply
    nested ElementI-compatible object, the dirty flag is set. The dirty flag
    is guaranteed to be called on changes to X, Y, Width, Height, and some others
    as long as the BaseElement type is embedded and BaseElement's CalculateStyle
    is not overriden or is directly called.
  * Upon the call of Client.Refresh(), if any children of RootWindow are found
    to have the dirty flag set, Render() is called on the RootWindow, which
    thereafter cascades downwards. When an Element is Rendered, the dirty flag
    is cleared.

Additionally, a Style system has been implemented that is used for storing
desired position, dimensions, colors, and otherwise. Some important values, such
as x, y, width, and height, are stored as precomputed/cached properties on the
element itself.