Decouple SDL2 from Client; Implement InputElement; Fix UI Style
SDL2 is not completely decoupled from the Client, but much of it has been
separated into UI/Instance. What remains is the custom RenderFunc calls
that can be defined in the Window or Element's custom Render member var,
however this could likely be removed or use a simplified SDL-like render
API that could be implemented in OpenGL or some other gfx library.

InputElement is functional. Much of this required restructuring UI event
handling to first go to a central UI state instance before iterating down
the Element tree.

UI Styling should, hopefully, but fixed, in terms of percentile positioning.
Style updating when the parent is a Window or an Element must be handled
differently, as Elements with a Window parent must position themselves relative
to the x,y within the Window texture, whilst Elements with an Element parent
must position themselves absolutely.