Use Channels for Element interactions; Add ContainerElement
MAJOR CHANGE: We are now using Channels for handling the adoption,
destruction, and updating of Elements. This allows elements to be
properly be accessed from goroutines without risking corruption or
crashes.

We have also split the WindowElement into a new ContainerElement. This
ContainerElement functions according to what was defined as a sub-window
before. It made more sense to split the logic into two distinct types
rather than having an excessive amount of conditionals that broke
readability.