Update OnKeyDown; add Game<->UI input channel
OnKeyDown now also provides a repeat argument to signify if the key is a
repeated value from the OS or not.

Additionally, we have a channel in Game that receives the UserInput
interface. This is used to process events that should exist in the
Client's goroutine, such as key bindings that trigger network sends or
world updating. At the moment it feels a little messy given how the UI
system is written, but it will suffice for now.