Optimize object rendering
We now have a separate objectImageIDs map that stores the object IDs to
their current rendered image ID. If there is a difference between what
the world object's frame ID is, we send an Image update to the UI element
and update our image ID map.

Ideally the SDL backend would actually create a global texture<->image
map and do some sort of comparison check for Image interfaces to
textures and use those internally. Unsure of how well this would work,
but it could be better.